+++ backend/tests/Designer.Tests/Controllers/PreviewController/GetRuleHandlerTests.cs
@@ -1,0 +1,26 @@
+﻿using System.Net;
+using System.Net.Http;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc.Testing;
+using Xunit;
+
+namespace Designer.Tests.Controllers.PreviewController
+{
+    public class GetRuleHandlerTests : PreviewControllerTestsBase<GetRuleHandlerTests>
+    {
+
+        public GetRuleHandlerTests(WebApplicationFactory<Altinn.Studio.Designer.Controllers.PreviewController> factory) : base(factory)
+        {
+        }
+
+        [Fact]
+        public async Task Get_RuleHandler_NoContent()
+        {
+            string dataPathWithData = $"{Org}/{App}/api/resource/RuleHandler.js";
+            using HttpRequestMessage httpRequestMessage = new(HttpMethod.Get, dataPathWithData);
+
+            using HttpResponseMessage response = await HttpClient.Value.SendAsync(httpRequestMessage);
+            Assert.Equal(HttpStatusCode.NoContent, response.StatusCode);
+        }
+    }
+}
